backend/requirements.txt
Django~=3.2.9
djangorestframework~=3.12.4
requests~=2.26.0
beautifulsoup4~=4.10.0
icalendar~=4.0.9
python-dateutil~=2.8.2
celery~=5.2.1